apps/mcp/pyproject.toml
[project]
name = "assetbridge-mcp"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.0",
    "sqlalchemy>=2.0",
    "pydantic-settings>=2.3",
    "python-dotenv>=1.0",
    "httpx>=0.27",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
